+++ lightstep/sdk/metric/exporters/otlp/otelcol/otlp.go
@@ -15,16 +15,22 @@
 package otelcol
 
 import (
+	"fmt"
+
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/internal"
+	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator/aggregation"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator/gauge"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator/histogram"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator/minmaxsumcount"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator/sum"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/data"
+	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/exemplar"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/number"
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/pmetric"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
 )
 
 func toTemporality(t aggregation.Temporality) pmetric.AggregationTemporality {
@@ -50,7 +56,7 @@ func copySumPoints(m pmetric.Metric, inM data.Instrument, mono bool) {
 
 		internal.CopyAttributes(dp.Attributes(), inP.Attributes)
 
-		switch t := inP.Aggregation.(type) {
+		switch t := unwrapExemplars(inP.Aggregation).(type) {
 		case *sum.MonotonicInt64:
 			dp.SetIntValue(number.ToInt64(t.Sum()))
 		case *sum.NonMonotonicInt64:
@@ -62,6 +68,8 @@ func copySumPoints(m pmetric.Metric, inM data.Instrument, mono bool) {
 		default:
 			panic("unhandled case")
 		}
+
+		copyExemplars(dp.Exemplars(), inP.Attributes, inM.Descriptor.NumberKind, inP.Exemplars)
 	}
 }
 
@@ -76,7 +84,7 @@ func copyGaugePoints(m pmetric.Metric, inM data.Instrument) {
 
 		internal.CopyAttributes(dp.Attributes(), inP.Attributes)
 
-		switch t := inP.Aggregation.(type) {
+		switch t := unwrapExemplars(inP.Aggregation).(type) {
 		case *gauge.Int64:
 			dp.SetIntValue(number.ToInt64(t.Gauge()))
 		case *gauge.Float64:
@@ -84,6 +92,8 @@ func copyGaugePoints(m pmetric.Metric, inM data.Instrument) {
 		default:
 			panic("unhandled case")
 		}
+
+		copyExemplars(dp.Exemplars(), inP.Attributes, inM.Descriptor.NumberKind, inP.Exemplars)
 	}
 }
 
@@ -99,7 +109,7 @@ func copyHistogramPoints(m pmetric.Metric, inM data.Instrument) {
 
 		internal.CopyAttributes(dp.Attributes(), inP.Attributes)
 
-		switch t := inP.Aggregation.(type) {
+		switch t := unwrapExemplars(inP.Aggregation).(type) {
 		case *histogram.Int64:
 			dp.SetSum(t.Sum().CoerceToFloat64(number.Int64Kind))
 			dp.SetCount(t.Count())
@@ -133,6 +143,8 @@ func copyHistogramPoints(m pmetric.Metric, inM data.Instrument) {
 		default:
 			panic("unhandled case")
 		}
+
+		copyExemplars(dp.Exemplars(), inP.Attributes, inM.Descriptor.NumberKind, inP.Exemplars)
 	}
 }
 
@@ -159,7 +171,7 @@ func copyMMSCPoints(m pmetric.Metric, inM data.Instrument) {
 
 		internal.CopyAttributes(dp.Attributes(), inP.Attributes)
 
-		switch t := inP.Aggregation.(type) {
+		switch t := unwrapExemplars(inP.Aggregation).(type) {
 		case *minmaxsumcount.Int64:
 			dp.SetSum(t.Sum().CoerceToFloat64(number.Int64Kind))
 			dp.SetCount(t.Count())
@@ -177,7 +189,16 @@ func copyMMSCPoints(m pmetric.Metric, inM data.Instrument) {
 		default:
 			panic("unhandled case")
 		}
+
+		copyExemplars(dp.Exemplars(), inP.Attributes, inM.Descriptor.NumberKind, inP.Exemplars)
 	}
+}
+
+func unwrapExemplars(agg aggregation.Aggregation) aggregation.Aggregation {
+	if unwr, ok := agg.(exemplar.Unwrapper); ok {
+		agg = unwr.Unwrap()
+	}
+	return agg
 }
 
 func (c *client) d2pd(in data.Metrics) pmetric.Metrics {
@@ -202,7 +223,9 @@ func (c *client) d2pd(in data.Metrics) pmetric.Metrics {
 			m.SetUnit(string(inM.Descriptor.Unit))
 			m.SetDescription(inM.Descriptor.Description)
 
-			switch inM.Points[0].Aggregation.(type) {
+			agg := unwrapExemplars(inM.Points[0].Aggregation)
+
+			switch agg.(type) {
 			case *sum.MonotonicInt64, *sum.MonotonicFloat64:
 				copySumPoints(m, inM, true)
 			case *sum.NonMonotonicInt64, *sum.NonMonotonicFloat64:
@@ -213,9 +236,74 @@ func (c *client) d2pd(in data.Metrics) pmetric.Metrics {
 				copyHistogramPoints(m, inM)
 			case *minmaxsumcount.Int64, *minmaxsumcount.Float64:
 				copyMMSCPoints(m, inM)
+			default:
+				otel.Handle(fmt.Errorf("unknown concrete aggregator type: %T", agg))
 			}
 		}
 	}
 
 	return out
+}
+
+func copyExemplars(dest pmetric.ExemplarSlice, attrs attribute.Set, nk number.Kind, src []aggregator.WeightedExemplarBits) {
+	for _, wex := range src {
+		ex := dest.AppendEmpty()
+		ex.SetTimestamp(pcommon.NewTimestampFromTime(wex.Time))
+		if nk == number.Int64Kind {
+			ex.SetIntValue(number.ToInt64(wex.Number))
+		} else {
+			ex.SetDoubleValue(number.ToFloat64(wex.Number))
+		}
+		if wex.Span != nil {
+			ex.SetTraceID(pcommon.TraceID(wex.Span.SpanContext().TraceID()))
+			ex.SetSpanID(pcommon.SpanID(wex.Span.SpanContext().SpanID()))
+		}
+		// The following calculation appears to be optimizable
+		// -- except it's not clear how.  We've computed an
+		// attribute set by a filter somewhere, and at that
+		// moment we know the filtered attributes.  However
+		// connecting these code points feels difficult, so
+		// for now we re-sort the original attributes, then
+		// iterate.
+		aset := attribute.NewSet(wex.Attributes...)
+
+		// attrs is a subset of aset
+		oiter := attrs.Iter()
+		fiter := aset.Iter()
+
+		o := oiter.Len()
+		f := fiter.Len()
+
+		oiter.Next()
+		fiter.Next()
+
+		// TL;DR wishing we could start from scratch with a
+		// redesigned attribute.Set.
+		for o > 0 && f > 0 {
+			okv := oiter.Attribute()
+			fkv := fiter.Attribute()
+
+			if fkv.Key == okv.Key {
+				o--
+				f--
+				oiter.Next()
+				fiter.Next()
+			} else {
+				internal.CopyAttribute(ex.FilteredAttributes(), fkv)
+				f--
+				fiter.Next()
+			}
+		}
+
+		for f > 0 {
+			fkv := fiter.Attribute()
+			internal.CopyAttribute(ex.FilteredAttributes(), fkv)
+			f--
+			fiter.Next()
+		}
+
+		if wex.Weight != 0 {
+			ex.FilteredAttributes().PutDouble("sample.weight", wex.Weight)
+		}
+	}
 }