+++ lightstep/sdk/metric/internal/test/test.go
@@ -20,9 +20,11 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator/aggregation"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator/gauge"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/data"
+	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/exemplar"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/number"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/sdkinstrument"
 	"go.opentelemetry.io/otel"
@@ -30,6 +32,8 @@ import (
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/resource"
+	"go.opentelemetry.io/otel/trace"
+	"go.opentelemetry.io/otel/trace/noop"
 )
 
 func Descriptor(name string, ik sdkinstrument.Kind, nk number.Kind) sdkinstrument.Descriptor {
@@ -48,6 +52,18 @@ func Point(start, end time.Time, agg aggregation.Aggregation, tempo aggregation.
 		Attributes:  attrs,
 		Aggregation: agg,
 		Temporality: tempo,
+	}
+}
+
+func PointEx(start, end time.Time, agg aggregation.Aggregation, tempo aggregation.Temporality, kvs []attribute.KeyValue, exs ...aggregator.WeightedExemplarBits) data.Point {
+	attrs := attribute.NewSet(kvs...)
+	return data.Point{
+		Start:       start,
+		End:         end,
+		Attributes:  attrs,
+		Aggregation: agg,
+		Temporality: tempo,
+		Exemplars:   exs,
 	}
 }
 
@@ -120,6 +136,14 @@ func RequireEqualPoints(t *testing.T, output []data.Point, expected ...data.Poin
 			out.End = exp.End
 		}
 
+		// Special case for testing exemplars: the exemplar aggregator is full of
+		// intermediate state, so we replace the aggregation with the
+		// underlying aggregation for the purposes of testing its correct
+		// value w/o testing sampler state.
+		if unwr, ok := out.Aggregation.(exemplar.Unwrapper); ok {
+			out.Aggregation = unwr.Unwrap()
+		}
+
 		if outig, ok := out.Aggregation.(*gauge.Int64); ok {
 			outig.SetSequenceForTesting()
 		}
@@ -128,7 +152,42 @@ func RequireEqualPoints(t *testing.T, output []data.Point, expected ...data.Poin
 		}
 	}
 
+	var cpyEx [][]aggregator.WeightedExemplarBits
+	var outEx [][]aggregator.WeightedExemplarBits
+
+	for i := range cpy {
+		if len(cpy[i].Exemplars) != 0 {
+			cpyEx = append(cpyEx, cpy[i].Exemplars)
+			cpy[i].Exemplars = nil
+		}
+	}
+	for i := range output {
+		if len(output[i].Exemplars) != 0 {
+			outEx = append(outEx, output[i].Exemplars)
+			output[i].Exemplars = nil
+		}
+	}
+
 	require.ElementsMatch(t, cpy, output)
+
+	require.Equal(t, len(cpyEx), len(outEx))
+
+	for i := range cpyEx {
+		if len(cpyEx[i]) != len(outEx[i]) {
+			require.ElementsMatch(t, cpyEx, outEx)
+			continue
+		}
+
+		// As with the above, we zero timestamps if the expectation
+		// is zero, to skip the timestamp test.
+		for j := range cpyEx[i] {
+			if cpyEx[i][j].Time.IsZero() {
+				outEx[i][j].Time = time.Time{}
+			}
+		}
+
+		require.ElementsMatch(t, cpyEx, outEx)
+	}
 }
 
 // RequireEqualMetrics checks that an output equals the expected
@@ -171,4 +230,24 @@ func OTelErrors() *[]error {
 		*errors = append(*errors, err)
 	}))
 	return errors
+}
+
+type FakeSpanData struct {
+	t, s byte
+	noop.Span
+}
+
+func FakeSpan(t, s byte) trace.Span {
+	return &FakeSpanData{
+		t: t,
+		s: s,
+	}
+}
+
+func (f *FakeSpanData) SpanContext() trace.SpanContext {
+	return trace.NewSpanContext(trace.SpanContextConfig{
+		TraceID:    [16]byte{f.t, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+		SpanID:     [8]byte{f.s, 0, 0, 0, 0, 0, 0, 0},
+		TraceFlags: 0x1,
+	})
 }